Include rust bindings instead of generating at compile time
includes both ittnotify and jitprofiling rust bindings instead
of building at compile time. This eliminates the need for
projects that use this to have to bring in bindgen. Also included
is a test to check to see that the bindings are up-to-date.

Patch also updates documentation location in Cargo.toml