ci: add initial GitHub workflow
This change adds GitHub workflows to build both the C library and Rust
crate that depends on it. Though it does not check that VTune
collection works correctly with the built library, this change closes
 #17 in expectation that this level of testing will be added later (it
requires an environment with hardware and applications that may be more
easy to control elsewhere?).