rust: modify Rust crate to support Windows
This change has several parts: first, it splits out ITT bindings by OS.
The ITT bindings have OS-specific variables (`ITT_OS`, `ITT_PLATFORM`)
that encode which OS the bindings are for. This change creates separate
binding files for each OS. The JIT bindings are also not OS-agnostic and
are split out to the same directories. This change also removes libc
constants present in the generated bindings and ignores Windows carriage
returns during testing. Now that the bindings are generated correctly
for each OS, this change builds the C libraries and links to them in the
`build.rs` script on Windows. As a side effect, this removes the
`force_32` feature--it was not applying the `-m32` flag correctly during
compilation--but the feature can be re-added if needed in the future.