rust: rename Rust crates directory (#50)
* rust: rename Rust crates directory

Previously, the Rust code lived under the `ittapi-rs` directory and was
published as a single crate: `ittapi-rs`. Now, with the split of the
crates into the high-level APIs, `ittapi`, and low-level interface,
`ittapi-sys`, this directory name does not fit as well. This change
makes it more clear to the casual visitor to this repository where the
Rust code is contained: `ittapi-rs` -> `rust`.