ci: add MinGW build for Rust crate (#61)
In https://github.com/bytecodealliance/wasmtime/pull/4003 we discovered that the `ittapi` C code uses the uncommon `strnlen_s` function. On older versions of MinGW (GitHub runners currently come with v8.1.0 pre-installed), this function was not available. Since v8.1.0, other versions have been published; this change re-installs MinGW with a recent version which includes `strnlen_s`. This commit has no code changes, just CI configuration to prove that the crate builds with the right MinGW version.

As a result of this, we opened https://github.com/actions/virtual-environments/issues/5530 to upgrade MinGW on all Windows GitHub runners. That upgrade may take some time so in the meantime the workaround is to use a new-ish version of MinGW. Note that this build issue is likely rare: it only really applies if someone builds the `ittapi` crate with `--target x86_64-pc-windows-gnu` AND with an old version of MinGW installed.

Co-authored-by: Andrew Brown <andrew.brown@intel.com>