rust: warn when misusing `Event`
This change enables clippy's `must_use` warning and adds `#[must_use]`
in all the places it expects us to. This means that Rust users of
`ittapi` will see a warning if they don't assign the result of these
functions to a variable. This is important because it can be a bit
confusing that `Event::start` creates an object that keeps track of the
event state; if that object is immediately dropped, the event will last
0 time.

Additionally, this change adds some documentation to `Event` to make it
clear how the result of `Event::start` should be used.