ci: only build 64-bit version
For some time, Ubuntu has been phasing out support for 32-bit packages
(e.g., [here]). `ittapi` still builds a 32-bit version of the library,
which requires `gcc-multilib` to be installed (it provides
`/usr/include/bits/libc-header-start.h` when compiling with `-m32`). But
some GitHub inconsistency is getting in the way: `gcc-multilib` fails to
install on the GitHub runner though it can be installed locally.

[here]: https://canonical.com/blog/statement-on-32-bit-i386-packages-for-ubuntu-19-10-and-20-04-lts

While we figure out the 32-bit support story for `ittapi`, this change
disables the 32-bit build by forcing it to be 64-bit exclusively. Once
we decide whether to (a) remove 32-bit support or (b) fix the GitHub
runners to allow 32-bit builds, we can remove the TODO.