rust: improve `StringHandle` test
PR #69 identified an issue with how the C library gives out string
handles: it only does so when `INTEL_LIBITTNOTIFY64` is present in the
environment. This change clarifies this in the test, though presumably
at some point the C library should not be handing out null pointers or
`StringHandle::new` should fail with a coherent error message.